electronic_shopping_webapp
A website that is meant for buying products online. USER-END:-This module is for users where they register themselves with the permission of the administrator. ADMINISTRATOR-END:-This module is for administrator where all the users and products are managed and verified Frontend: HTML, CSS, JavaScript, JQuery Backend: SQL Server, Servlet, JSP, Hibernate Allows the user to add any comment or complaint or any kind of feedback. Helps the administrator to show all the registered users.